+++ src/arpes/fits/fit_models/__init__.py
@@ -10,7 +10,6 @@ from .fermi_edge import (
     AffineBroadenedFD,
     BandEdgeBGModel,
     BandEdgeBModel,
-    FermiDiracAffGaussModel,
     FermiDiracModel,
     FermiLorentzianModel,
     GStepBModel,